exam

+++ src/cart/cartSlice/index.js
@@ -1,4 +1,5 @@
 import { createSlice } from "@reduxjs/toolkit";
+import { message } from "antd";
 
 const initialState = {
   products: JSON.parse(localStorage.getItem("cart")) ?? [],
@@ -9,19 +10,30 @@ const cartSlice = createSlice({
   initialState,
   reducers: {
     addToCart(state, action) {
-      state.products = [...state.products, action.payload];
-      console.log(state.products);
-      localStorage.setItem("cart", JSON.stringify(state.products));
+      const existingItem = state.products.find(
+        (item) => item.id === action.payload.id
+      );
+      if (existingItem) {
+        message.warning("Item already in cart");
+      } else {
+        state.products = [...state.products, action.payload];
+        message.success("Item added to cart");
+        localStorage.setItem("cart", JSON.stringify(state.products));
+      }
     },
     deleteToCart(state, action) {
-      console.log(action.payload);
       state.products = state.products.filter(
         (item) => item.id !== action.payload.id
       );
+      localStorage.setItem("cart", JSON.stringify(state.products));
+      message.success("Item deleted from cart");
+    },
+    amountCart(state, action) {
+      state.products = action.payload;
       localStorage.setItem("cart", JSON.stringify(state.products));
     },
   },
 });
 
-export const { addToCart, deleteToCart } = cartSlice.actions;
+export const { addToCart, deleteToCart, amountCart } = cartSlice.actions;
 export default cartSlice.reducer;
